feat: refactor IAM content list API and mainserver request handling with improved timeout management
