+++ apps/sva-studio-react/src/lib/iam-content-list-api.server.ts
@@ -1,13 +1,4 @@
-import type {
-  ApiErrorCode,
-  ApiErrorResponse,
-  ApiListResponse,
-  ApiPagination,
-  ContentJsonValue,
-  IamContentAccessSummary,
-  IamContentListItem,
-  IamContentListQuery,
-} from '@sva/core';
+import type { ApiListResponse, IamContentListItem, IamContentListQuery } from '@sva/core';
 import {
   authorizeContentPrimitiveForUser,
   type AuthenticatedRequestContext,
@@ -18,103 +9,22 @@ import {
   listSvaMainserverNews,
   listSvaMainserverPoi,
 } from '@sva/sva-mainserver/server';
-import type {
-  SvaMainserverEventItem,
-  SvaMainserverNewsItem,
-  SvaMainserverPoiItem,
-} from '@sva/sva-mainserver';
 import { getWorkspaceContext } from '@sva/server-runtime';
 
 import { listContentsHandler } from '@sva/auth-runtime/runtime-routes';
-
-const MAIN_SERVER_CONTENT_TYPES = new Set([
-  'news.article',
-  'events.event-record',
-  'poi.point-of-interest',
-]);
-const MAINSERVER_FETCH_PAGE_SIZE = 100;
-const API_ERROR_CODES = new Set<ApiErrorCode>([
-  'unauthorized',
-  'forbidden',
-  'not_found',
-  'invalid_request',
-  'invalid_instance_id',
-  'invalid_organization_id',
-  'organization_inactive',
-  'rate_limited',
-  'csrf_validation_failed',
-  'idempotency_key_required',
-  'idempotency_key_reuse',
-  'idempotency_in_progress',
-  'keycloak_unavailable',
-  'tenant_auth_client_secret_missing',
-  'tenant_admin_client_not_configured',
-  'tenant_admin_client_secret_missing',
-  'encryption_not_configured',
-  'database_unavailable',
-  'last_admin_protection',
-  'self_protection',
-  'feature_disabled',
-  'conflict',
-  'legal_acceptance_required',
-  'reauth_required',
-  'internal_error',
-]);
-
-const isMainserverContentType = (value: string): boolean => MAIN_SERVER_CONTENT_TYPES.has(value);
-const normalizeApiErrorCode = (value: unknown): ApiErrorCode =>
-  typeof value === 'string' && API_ERROR_CODES.has(value as ApiErrorCode)
-    ? (value as ApiErrorCode)
-    : 'internal_error';
-
-const readContentListQuery = (request: Request): IamContentListQuery => {
-  const url = new URL(request.url);
-  const page = Math.max(1, Number.parseInt(url.searchParams.get('page') ?? '1', 10) || 1);
-  const pageSize = Math.max(1, Number.parseInt(url.searchParams.get('pageSize') ?? '25', 10) || 25);
-  const q = url.searchParams.get('q')?.trim() || undefined;
-  const typeValue = url.searchParams.get('type')?.trim();
-  const statusValue = url.searchParams.get('status')?.trim();
-  const sortByValue = url.searchParams.get('sortBy')?.trim();
-  const sortDirectionValue = url.searchParams.get('sortDirection')?.trim();
-  const visibleTypes = url.searchParams
-    .getAll('visibleType')
-    .map((value) => value.trim())
-    .filter((value) => value.length > 0);
-
-  return {
-    page,
-    pageSize,
-    ...(q ? { q } : {}),
-    ...(typeValue && typeValue !== 'all' ? { type: typeValue } : {}),
-    ...(statusValue === 'draft' ||
-    statusValue === 'in_review' ||
-    statusValue === 'approved' ||
-    statusValue === 'published' ||
-    statusValue === 'archived'
-      ? { status: statusValue }
-      : {}),
-    ...(visibleTypes.length > 0 ? { visibleTypes } : {}),
-    sortBy:
-      sortByValue === 'title' ||
-      sortByValue === 'contentType' ||
-      sortByValue === 'status' ||
-      sortByValue === 'updatedAt'
-        ? sortByValue
-        : 'updatedAt',
-    sortDirection: sortDirectionValue === 'asc' || sortDirectionValue === 'desc' ? sortDirectionValue : 'desc',
-  };
-};
-
-const partitionRequestedTypes = (query: IamContentListQuery): {
-  readonly localTypes: readonly string[];
-  readonly mainserverTypes: readonly string[];
-} => {
-  const requestedTypes = query.type ? [query.type] : (query.visibleTypes ?? []);
-  return {
-    mainserverTypes: requestedTypes.filter(isMainserverContentType),
-    localTypes: requestedTypes.filter((value) => !isMainserverContentType(value)),
-  };
-};
+import {
+  createListErrorResponse,
+  createListResponse,
+  createListSubrequest,
+  filterItems,
+  MAINSERVER_FETCH_PAGE_SIZE,
+  normalizeApiErrorCode,
+  paginateItems,
+  partitionRequestedTypes,
+  readContentListQuery,
+  sortItems,
+} from './iam-content-list-api.shared.js';
+import { mapEventItem, mapNewsItem, mapPoiItem } from './iam-content-list-mainserver.js';
 
 const readPermissionActions = (
   permissions: readonly Readonly<{
@@ -125,276 +35,6 @@ const readPermissionActions = (
   permissions
     .filter((permission) => permission.effect !== 'deny')
     .map((permission) => permission.action);
-
-const deriveUpdateAction = (contentType: string): string | null => {
-  const namespace = contentType.split('.')[0]?.trim();
-  return namespace ? `${namespace}.update` : null;
-};
-
-const deriveCreateAction = (contentType: string): string | null => {
-  const namespace = contentType.split('.')[0]?.trim();
-  return namespace ? `${namespace}.create` : null;
-};
-
-const createMainserverItemAccess = (
-  contentType: string,
-  permissionActions: readonly string[]
-): IamContentAccessSummary => {
-  const updateAction = deriveUpdateAction(contentType);
-  const createAction = deriveCreateAction(contentType);
-  const canUpdate = updateAction ? permissionActions.includes(updateAction) : false;
-  const canCreate = createAction ? permissionActions.includes(createAction) : false;
-
-  return canUpdate
-    ? {
-        state: 'editable',
-        canRead: true,
-        canCreate,
-        canUpdate: true,
-        organizationIds: [],
-        sourceKinds: [],
-      }
-    : {
-        state: 'read_only',
-        canRead: true,
-        canCreate,
-        canUpdate: false,
-        reasonCode: 'content_update_missing',
-        organizationIds: [],
-        sourceKinds: [],
-      };
-};
-
-const toContentJsonValue = (value: unknown): ContentJsonValue =>
-  JSON.parse(JSON.stringify(value ?? null)) as ContentJsonValue;
-
-const normalizeTitle = (value: string | undefined, fallback: string): string => {
-  const normalized = value?.trim();
-  return normalized && normalized.length > 0 ? normalized : fallback;
-};
-
-const resolveNewsTitle = (item: SvaMainserverNewsItem): string =>
-  normalizeTitle(item.title, normalizeTitle(item.contentBlocks?.[0]?.title, item.id));
-
-const mapNewsItem = (
-  item: SvaMainserverNewsItem,
-  instanceId: string,
-  permissionActions: readonly string[]
-): IamContentListItem => ({
-  id: item.id,
-  instanceId,
-  contentType: item.contentType,
-  title: resolveNewsTitle(item),
-  createdAt: item.createdAt,
-  createdBy: item.author,
-  updatedAt: item.updatedAt,
-  updatedBy: item.author,
-  author: item.author,
-  payload: toContentJsonValue(item.payload),
-  status: 'published',
-  validationState: 'valid',
-  historyRef: `mainserver:news:${item.id}`,
-  publishedAt: item.publishedAt,
-  access: createMainserverItemAccess(item.contentType, permissionActions),
-});
-
-const mapEventItem = (
-  item: SvaMainserverEventItem,
-  instanceId: string,
-  permissionActions: readonly string[]
-): IamContentListItem => ({
-  id: item.id,
-  instanceId,
-  contentType: item.contentType,
-  title: normalizeTitle(item.title, item.id),
-  createdAt: item.createdAt,
-  createdBy: 'mainserver',
-  updatedAt: item.updatedAt,
-  updatedBy: 'mainserver',
-  author: 'mainserver',
-  payload: toContentJsonValue({
-    description: item.description,
-    categoryName: item.categoryName,
-    dates: item.dates,
-    addresses: item.addresses,
-    contacts: item.contacts,
-    urls: item.urls,
-    tags: item.tags,
-  }),
-  status: 'published',
-  validationState: 'valid',
-  historyRef: `mainserver:events:${item.id}`,
-  access: createMainserverItemAccess(item.contentType, permissionActions),
-});
-
-const mapPoiItem = (
-  item: SvaMainserverPoiItem,
-  instanceId: string,
-  permissionActions: readonly string[]
-): IamContentListItem => ({
-  id: item.id,
-  instanceId,
-  contentType: item.contentType,
-  title: normalizeTitle(item.name, item.id),
-  createdAt: item.createdAt,
-  createdBy: 'mainserver',
-  updatedAt: item.updatedAt,
-  updatedBy: 'mainserver',
-  author: 'mainserver',
-  payload: toContentJsonValue({
-    description: item.description,
-    mobileDescription: item.mobileDescription,
-    active: item.active,
-    categoryName: item.categoryName,
-    payload: item.payload,
-    addresses: item.addresses,
-    contact: item.contact,
-    openingHours: item.openingHours,
-    webUrls: item.webUrls,
-    tags: item.tags,
-  }),
-  status: 'published',
-  validationState: 'valid',
-  historyRef: `mainserver:poi:${item.id}`,
-  access: createMainserverItemAccess(item.contentType, permissionActions),
-});
-
-const toSearchableText = (item: IamContentListItem): string =>
-  [item.title, item.contentType, item.author, JSON.stringify(item.payload)]
-    .join(' ')
-    .toLowerCase();
-
-const filterItems = (items: readonly IamContentListItem[], query: IamContentListQuery): readonly IamContentListItem[] => {
-  const normalizedSearch = query.q?.trim().toLowerCase();
-
-  return items.filter((item) => {
-    if (query.type && item.contentType !== query.type) {
-      return false;
-    }
-    if (query.status && item.status !== query.status) {
-      return false;
-    }
-    if (normalizedSearch && !toSearchableText(item).includes(normalizedSearch)) {
-      return false;
-    }
-    return true;
-  });
-};
-
-const compareItemsBySortField = (
-  left: IamContentListItem,
-  right: IamContentListItem,
-  sortBy: IamContentListQuery['sortBy'],
-  collator: Intl.Collator
-): number => {
-  switch (sortBy) {
-    case 'contentType':
-      return collator.compare(left.contentType, right.contentType);
-    case 'title':
-      return collator.compare(left.title, right.title);
-    case 'status':
-      return collator.compare(left.status, right.status);
-    default:
-      return collator.compare(left.updatedAt, right.updatedAt);
-  }
-};
-
-const sortItems = (
-  items: readonly IamContentListItem[],
-  sortBy: IamContentListQuery['sortBy'],
-  sortDirection: IamContentListQuery['sortDirection']
-): readonly IamContentListItem[] => {
-  const direction = sortDirection === 'asc' ? 1 : -1;
-  const collator = new Intl.Collator('de', { sensitivity: 'base', numeric: true });
-
-  return [...items].sort((left, right) => {
-    const result = compareItemsBySortField(left, right, sortBy, collator);
-    if (result !== 0) {
-      return result * direction;
-    }
-    return collator.compare(left.id, right.id) * direction;
-  });
-};
-
-const paginateItems = (items: readonly IamContentListItem[], page: number, pageSize: number): readonly IamContentListItem[] => {
-  const offset = Math.max(0, (page - 1) * pageSize);
-  return items.slice(offset, offset + pageSize);
-};
-
-const createListErrorResponse = (
-  status: number,
-  code: ApiErrorCode,
-  message: string,
-  requestId?: string
-): Response =>
-  new Response(
-    JSON.stringify({
-      error: {
-        code,
-        message,
-      },
-      ...(requestId ? { requestId } : {}),
-    } satisfies ApiErrorResponse),
-    {
-      status,
-      headers: { 'Content-Type': 'application/json' },
-    }
-  );
-
-const createListResponse = (
-  items: readonly IamContentListItem[],
-  pagination: ApiPagination,
-  requestId?: string
-): Response =>
-  new Response(
-    JSON.stringify({
-      data: items,
-      pagination,
-      ...(requestId ? { requestId } : {}),
-    } satisfies ApiListResponse<IamContentListItem>),
-    {
-      status: 200,
-      headers: { 'Content-Type': 'application/json' },
-    }
-  );
-
-const appendQueryValues = (searchParams: URLSearchParams, query: IamContentListQuery): void => {
-  searchParams.set('page', String(query.page));
-  searchParams.set('pageSize', String(query.pageSize));
-
-  if (query.q) {
-    searchParams.set('q', query.q);
-  }
-  if (query.type) {
-    searchParams.set('type', query.type);
-  }
-  if (query.status) {
-    searchParams.set('status', query.status);
-  }
-  if (query.sortBy) {
-    searchParams.set('sortBy', query.sortBy);
-  }
-  if (query.sortDirection) {
-    searchParams.set('sortDirection', query.sortDirection);
-  }
-  for (const visibleType of query.visibleTypes ?? []) {
-    searchParams.append('visibleType', visibleType);
-  }
-};
-
-const createListSubrequest = (
-  request: Request,
-  query: IamContentListQuery
-): Request => {
-  const url = new URL(request.url);
-  url.search = '';
-  appendQueryValues(url.searchParams, query);
-
-  return new Request(url.toString(), {
-    method: 'GET',
-    headers: request.headers,
-  });
-};
 
 const fetchAllPages = async <TItem>(
   loadPage: (query: { readonly page: number; readonly pageSize: number }) => Promise<{
